Fix compile warning
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@11719 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/modules/m_spanningtree/hmac.cpp
@@ -102,7 +102,7 @@ std::string TreeSocket::RandString(unsigned int ilength)
 
 	if (f >= 0)
 	{
-		if (read(f, randombuf, ilength) < ilength)
+		if (read(f, randombuf, ilength) < (int)ilength)
 			ServerInstance->Logs->Log("m_spanningtree", DEFAULT, "Entropy source has gone predictable (did not return enough data)");
 		close(f);
 	}
